EDX-4552-Resolve Comments
I have resolved the comments provided by Steve. Regarding the comment about removing the hard-coded "tleData.txt", it is difficult to change because the file in OneSGP4_Example/Program.cs is used only for internal purposes and testing. Also, OneSGP4_Example/Program.cs is a console application, so it does not support a UI to receive the file name from a browse button.